refactor: split types/types.go — JSON conversion into separate file
Extract JSON type conversion functions into json_convert.go; split matching
tests into json_convert_test.go; relocate TIMESTAMP converter tests to
converter_test.go where the tested functions reside.
